task#427: unify user-api
