+++ Cargo.toml
@@ -147,8 +147,6 @@ smoking-tests-macros = { path = "src/aws-tests/src/smoking-tests/src/smoking-tes
 staging-tests-macros = { path = "src/aws-tests/src/staging-tests/src/staging-tests-macros" }
 user = { path = "src/user" }
 user-api = { path = "src/user-api" }
-user-api-get-account = { path = "src/user-api/src/user-api-get-account" }
-user-api-patch-account = { path = "src/user-api/src/user-api-patch-account" }
 user-lambda = { path = "src/user-lambda" }
 cognito = { path = "src/cognito" }
 user-lambda-fanout-update-watchlist = { path = "src/user-lambda/src/user-lambda-fanout-update-watchlist" }
