+++ src/user-api/Cargo.toml
@@ -4,5 +4,29 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-user-api-get-account = { workspace = true }
-user-api-patch-account = { workspace = true }
+common = { workspace = true, features = ["api"] }
+user = { workspace = true, features = ["dynamodb", "service", "data"] }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serde-email = { workspace = true }
+serde_fields = { workspace = true }
+serde_dynamo = { workspace = true, features = ["aws-sdk-dynamodb+1"] }
+aws-sdk-dynamodb = { workspace = true }
+aws-config = { workspace = true }
+lambda_runtime = { workspace = true }
+aws_lambda_events = { workspace = true, features = ["apigw"] }
+time = { workspace = true }
+tokio = { workspace = true, features = ["macros"] }
+async-trait = { workspace = true }
+mockall = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true, features = ["json"] }
+http = { workspace = true }
+
+[dev-dependencies]
+test-api = { workspace = true, features = ["api-gateway", "dynamodb"] }
+serial_test = { workspace = true }
+fake = { workspace = true }
+user = { workspace = true, features = ["test-data"] }
+common = { workspace = true, features = ["test-data"] }
