+++ src/user-api/src/get.rs
@@ -1,38 +1,11 @@
 use aws_lambda_events::apigw::{ApiGatewayV2httpRequest, ApiGatewayV2httpResponse};
 use common::api::api_gateway_v2_http_response_builder::ApiGatewayV2HttpResponseBuilder;
-use common::api::error::{ApiError, log_api_error};
+use common::api::error::ApiError;
 use common::user_id::api::extract_user_id_request_context;
 use lambda_runtime::LambdaEvent;
 use user::data::get_user_data::GetUserAccountData;
 use user::service::user_service::UserService;
 
-#[tracing::instrument(
-    skip(event, service),
-    fields(
-        requestId = %event.context.request_id,
-        method = event.payload.request_context.http.method.as_str(),
-        path = &event.payload.raw_path.as_deref().unwrap_or("NULL"),
-        query = &event.payload.raw_query_string.as_deref().unwrap_or("NULL"),
-        body = &event.payload.body.as_deref().unwrap_or("NULL"),
-        ip = &event.payload.request_context.http.source_ip.as_deref().unwrap_or("NULL"),
-        userAgent = &event.payload.request_context.http.user_agent.as_deref().unwrap_or("NULL"),
-        userId = tracing::field::Empty,
-    )
-)]
-pub async fn handler(
-    event: LambdaEvent<ApiGatewayV2httpRequest>,
-    service: &impl UserService,
-) -> Result<ApiGatewayV2httpResponse, lambda_runtime::Error> {
-    match handle(event, service).await {
-        Ok(response) => Ok(response),
-        Err(err) => {
-            log_api_error(&err);
-            Ok(ApiGatewayV2httpResponse::from(err))
-        }
-    }
-}
-
-// GET /api/v1/me/account
 pub async fn handle(
     event: LambdaEvent<ApiGatewayV2httpRequest>,
     service: &impl UserService,
@@ -50,7 +23,7 @@ pub async fn handle(
 
 #[cfg(test)]
 mod tests {
-    use crate::handler;
+    use crate::handle;
     use common::user_id::UserId;
     use fake::{Fake, Faker};
     use http::header::LAST_MODIFIED;
@@ -74,11 +47,12 @@ mod tests {
         let lambda_event = LambdaEvent {
             payload: ApiGatewayV2httpRequestProxy::builder()
                 .http_method(http::Method::GET)
+                .route_key("GET /api/v1/me/account")
                 .jwt_claim("sub", UserId::new())
                 .build(),
             context: Default::default(),
         };
-        let response = handler(lambda_event, &service).await.unwrap();
+        let response = handle(lambda_event, &service).await.unwrap();
         assert_eq!(200, response.status_code);
         assert_eq!(
             "Wed, 01 Jan 2020 00:00:00 GMT",
@@ -93,12 +67,13 @@ mod tests {
         let lambda_event = LambdaEvent {
             payload: ApiGatewayV2httpRequestProxy::builder()
                 .http_method(http::Method::GET)
+                .route_key("GET /api/v1/me/account")
                 .build(),
             context: Default::default(),
         };
 
-        let response = handler(lambda_event, &service).await.unwrap();
-        assert_eq!(401, response.status_code);
+        let response = handle(lambda_event, &service).await.unwrap_err();
+        assert_eq!(401, response.status);
     }
 
     #[tokio::test]
@@ -106,6 +81,7 @@ mod tests {
         let lambda_event = LambdaEvent {
             payload: ApiGatewayV2httpRequestProxy::builder()
                 .http_method(http::Method::GET)
+                .route_key("GET /api/v1/me/account")
                 .jwt_claim("sub", UserId::new())
                 .build(),
             context: Default::default(),
@@ -117,7 +93,7 @@ mod tests {
             Box::pin(async move { Err(UserServiceError::UserNotFound(user_id)) })
         });
 
-        let response = handler(lambda_event, &service).await.unwrap();
-        assert_eq!(404, response.status_code);
+        let response = handle(lambda_event, &service).await.unwrap_err();
+        assert_eq!(404, response.status);
     }
 }
