+++ src/user-api/tests/get.rs
@@ -6,7 +6,7 @@ use user::{
     dynamodb::repository::UserDynamoDbRepositoryImpl,
     service::user_service::{UserService, UserServiceImpl},
 };
-use user_api_get_account::handler;
+use user_api::handler;
 
 #[localstack_test(services = [DynamoDB()])]
 async fn should_200_respond_user_when_exists() {
@@ -17,6 +17,7 @@ async fn should_200_respond_user_when_exists() {
     let lambda_event = LambdaEvent {
         payload: ApiGatewayV2httpRequestProxy::builder()
             .http_method(http::Method::GET)
+            .route_key("GET /api/v1/me/account")
             .jwt_claim("sub", user.user_id)
             .build(),
         context: Default::default(),
